+++ src/test/scala/scorex/lagonaki/integration/api/UtilsAPISpecification.scala
@@ -2,7 +2,7 @@ package scorex.lagonaki.integration.api
 
 import org.scalatest.{FunSuite, Matchers}
 import scorex.crypto.encode.Base58
-import scorex.crypto.hash.SecureCryptographicHash
+import scorex.crypto.hash.{FastCryptographicHash, SecureCryptographicHash}
 
 import scala.util.Random
 
@@ -10,11 +10,18 @@ class UtilsAPISpecification extends FunSuite with Matchers {
 
   import scorex.lagonaki.TestingCommons._
 
-  test("/utils/hash API route") {
+  test("/utils/hash/secure API route") {
     val msg = "test"
-    val resp = postRequest("/utils/hash", Map("message" -> msg))
+    val resp = postRequest("/utils/hash/secure", Map("message" -> msg))
     (resp \ "message").as[String] shouldBe msg
     (resp \ "hash").as[String] shouldBe Base58.encode(SecureCryptographicHash(msg))
+  }
+
+  test("/utils/hash/fast API route") {
+    val msg = "test"
+    val resp = postRequest("/utils/hash/fast", Map("message" -> msg))
+    (resp \ "message").as[String] shouldBe msg
+    (resp \ "hash").as[String] shouldBe Base58.encode(FastCryptographicHash(msg))
   }
 
   test("/utils/seed API route") {